+++ sc/sum1error.c
@@ -1,4 +1,4 @@
-int a;
+byte a;
 void main() {
 	a = 1+2;
 	b = a * 2;
